+++ aws-api-appsync/src/test/java/com/amplifyframework/api/aws/appsync/SelectionSetTest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.api.aws;
+package com.amplifyframework.api.aws.appsync;
 
 import com.amplifyframework.AmplifyException;
 import com.amplifyframework.api.graphql.QueryType;
