+++ aws-api-appsync/src/test/resources/selection-set-post.txt
